refactor: share timeline route frame
